Relocate Claude subagent classes and improve multi-provider docs
Move Claude subagent classes from tools.task.subagent.claude to
tools.task.claude, removing the intermediate subagent package.
Add javadoc for doFindChatClientBuilder explaining the provider:model
resolution logic. Update TaskTools.md to document multi-provider
routing with any Spring AI supported LLM provider. Update migration
guide with the new Claude package paths.
